Make use of the `japicmp` plugin to check for binary compatibility
This commit introduces the use of the JApiCmp plugin to check for binary compatibility between releases of Gradle.
The binary compatibility checks now generate a report per project, based on the last release version. Accepted
changes need to be added to the `distributions/src/changes/changes-since-x.y.txt` file.

The build will fail if a non accepted, binary breaking change, is introduced.