Use a single binary compatibility check task
This commit addresses a concern with regards to refactoring between modules. With individual reports, some errors would occur because
we moved classes from one module to the other, while, in practice, all are shipped together (as of now). This means that you cannot,
in practice, consider all artifacts as separate. They form a whole which is "the Gradle API", for good or for bad.